6/22_1 FlashAttention and Triton

+++ cs336_systems/flash_attention.py
@@ -70,7 +70,7 @@ def flash_fwd_kernel(
         K_j = tl.load(K_blk, boundary_check=(0,))  # SRAM으로 이동
         V_j = tl.load(V_blk, boundary_check=(0,))  # SRAM으로 이동
 
-        S_ij = tl.dot(Q_i, tl.trans(K_j), out_dtype=tl.float32) * scale  # (B_q, B_k)
+        S_ij = tl.dot(Q_i, tl.trans(K_j).to(tl.float32), out_dtype=tl.float32) * scale  # (B_q, B_k)
 
         if is_causal:
             q_idx = (q_start + tl.arange(0, B_q))[:, None] # (B_q, 1)
